app/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
attrs==23.1.0
bcrypt==4.0.1
black==23.11.0
blinker==1.7.0
boto3==1.29.0
botocore==1.32.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dill==0.3.7
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
langdetect==1.0.9
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
pdf2image==1.16.3
Pillow==10.1.0
platformdirs==4.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydotenv==0.0.7
pylint==3.0.2
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
referencing==0.31.0
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
Werkzeug==3.0.1
